GC: Bitset mark phase
The basic idea here is that we set a "finger" in the heap
(next_mark_index) during garbage collection.  Anything below the
finger is recursively marked as always.  Anything above the finger is
marked but we don't recurse.  Initially the finger points to the start
of the heap, but we gradually move it up over time.  Each time we move
it, we unmark the next live object we see and re-mark it (this
replaces the recursion).  So instead of building a large recursion
stack, in the best case we use the mark set itself to do the job.

I instrumented the GC first to discover which direction is most
beneficial, and it looks like it's a wash.  But I assume a lot of the
backward pointers are to the original program, the constant table, and
so forth and thus unlikely to change.

Overall this is a wash for GC time during bootstrap, but it's a *very*
interesting jumping-off point for more experiments.