Add tupling combinator T.
T takes a single numeric argument n,
followed by n field arguments and a continuation.

Its rewrite rule is:
  T i x1 ... xi k = k x1 ... xi

The goal here is to save approximately 2x heap footprint for every
tuple / ADT disjunct with more than 3 fields (where we break even).
This is called out in EncodeData as an area for exploration.

With 25 runs of bootstrap timings, it looks like this represents a
3.7-4.6% speedup depending on how you do it:
		user+sys	stddev	user	stddev	sys	stddev	user+sy speedup	user		sys
master	        cached	10.809	0.185	10.670	0.168	0.139	0.021	0.00%		0.00%		0.00%
master	        full	17.636	0.274	17.469	0.256	0.167	0.027	0.00%		0.00%		0.00%
master	        libs	15.910	0.290	15.733	0.268	0.177	0.028	0.00%		0.00%		0.00%
TupleCombinator	cached	10.404	0.186	10.266	0.172	0.138	0.018	3.75%		3.79%		0.95%
TupleCombinator	full	16.877	0.253	16.711	0.235	0.166	0.026	4.30%		4.34%		0.61%
TupleCombinator	libs	15.169	0.198	15.000	0.190	0.169	0.015	4.66%		4.66%		4.52%

I think in part this is because it's saving a GC on each run.  Note
that a similar version that introduces combinators in improve.hs
doesn't work quite as well -- it looks like we're better off
preserving full laziness for functions and reserving the new
combinator specifically for data.

It uses quite a bit less heap at peak.
 % /tmp/mhs-TupleCombinator +RTS -v -H4M -RTS -s -isrc -imhs -ipaths MicroHs.Main
677 lines/s
        126,012 combinator file size
         95,010 cells at start
      4,000,000 cells heap size (64,000,000 bytes)
  1,783,428,204 cells allocated (368.6 Mbyte/s)
            899 GCs
      3,876,485 max cells used
  4,459,293,287 reductions (57.6 Mred/s)
            548 array alloc
            547 array free
            544 foreign alloc
            544 foreign free
        623,101 bytestring alloc (max 53,222)
      5,755,589 bytestring alloc bytes (max 1,113,195)
        623,101 bytestring free
          77.42s total expired time
          60.02s total gc time (59.56 + 0.45)
 GC reductions A=2331, K=0, I=0, int=906711 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=0, R=0

Whereas on master:
% bin/mhs +RTS -v -H4M -RTS -s -isrc -imhs -ipaths MicroHs.Main
ERR: heap exhausted

It requires about 1.5x the heap to run current master vs this change:
% /tmp/mhs-master +RTS -v -H6M -RTS -isrc -imhs -ipaths MicroHs.Main
        126,182 combinator file size
         95,561 cells at start
      6,000,000 cells heap size (96,000,000 bytes)
  1,677,824,482 cells allocated (664.0 Mbyte/s)
            394 GCs
      3,958,585 max cells used
  4,235,862,731 reductions (104.8 Mred/s)
            419 array alloc
            418 array free
            415 foreign alloc
            415 foreign free
        623,186 bytestring alloc (max 55,008)
      5,755,759 bytestring alloc bytes (max 1,142,615)
        623,186 bytestring free
          40.43s total expired time
          25.21s total gc time (25.02 + 0.19)
 GC reductions A=2325, K=0, I=0, int=871594 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=0, R=0

It also yields dumps that are easier to read.

This has got to be a version bump of some kind.  It's backward-
compatible but definitely not forward compatible.