Add tupling combinator T.
T takes a single numeric argument n,
followed by n field arguments and a continuation.

Its rewrite rule is:
  T i x1 ... xi k = k x1 ... xi

The goal here is to save approximately 2x heap footprint for every
tuple / ADT disjunct with more than 3 fields.  This is called out in
EncodeData as an area for exploration.

With 25 runs of bootstrap timings, it looks like this represents a
2-3.5% speedup depending on how you do it:
		        user+sy	stddev	user	stddev	sys	stddev	user+sys	user		sys
TupleCombinator	cached	11.426	0.185	11.310	0.186	0.116	0.009	2.60%		2.65%		-2.25%
TupleCombinator	full	18.360	0.318	18.223	0.318	0.138	0.013	2.85%		2.88%		-1.37%
TupleCombinator	libs	16.594	0.283	16.453	0.286	0.141	0.008	3.62%		3.64%		1.23%
master	        cached	11.731	0.167	11.617	0.167	0.114	0.010	0.00%		0.00%		0.00%
master	        full	18.898	0.250	18.763	0.255	0.136	0.011	0.00%		0.00%		0.00%
master	        libs	17.217	0.274	17.074	0.277	0.143	0.011	0.00%		0.00%		0.00%

More important, it uses quite a bit less heap.
% bin/mhs +RTS -v -H4M -RTS -s -isrc -imhs -ipaths MicroHs.Main
730 lines/s
        435,361 combinator file size
         94,863 cells at start
      4,000,000 cells heap size (64,000,000 bytes)
  1,768,601,193 cells allocated (387.9 Mbyte/s)
            879 GCs
      3,832,958 max cells used
  4,610,169,444 reductions (63.2 Mred/s)
            552 array alloc
            551 array free
            548 foreign alloc
            548 foreign free
        622,837 bytestring alloc (max 53,434)
      5,754,418 bytestring alloc bytes (max 1,114,306)
        622,837 bytestring free
          72.95s total expired time
          55.27s total gc time (54.83 + 0.43)
 GC reductions A=2329, K=0, I=0, int=877438 flip=0
 special reductions B'=0 K4=0 K3=0 K2=0 C'B=0, Z=0, R=0

Whereas on master:
% bin/mhs +RTS -v -H4M -RTS -s -isrc -imhs -ipaths MicroHs.Main
ERR: heap exhausted

It requires about 1.5x the heap to run current master vs this change.
It also yields a much smaller compiler, and dumps are easier to read.

But this has got to be a version bump of some kind.  It's backward-compatible
but definitely not forward compatible.