Add iterated B combinators B2...B6.
B3 K, for example, means B (B (B K)).  This is common in target code.
It's important to preserve full laziness here, though, so we should think
of these as largely a compression mechanism for the repeated combinators.
If we do a bulk rewrite for full-arity applications, programs slow to a
crawl and allocate and GC like crazy.  It would be great if we could
identify exactly the cases where full laziness helps rather than costing;
for now, no such luck.