Add C2 combinator
This is a bit more conservative than the indexed tupling combinator I added in the
TupleCombinator branch.  It simply grabs pairs of arguments at a go.  In principle
it applies in more places, and we introduce it in Abstract (which we did not do
with the tupling combinator as it loses full laziness).

Speedups are at best minuscule:
		        user+sy	stddev	user	stddev	sys	stddev	user+sy speedup	user		sys
master	        cached	10.728	0.160	10.613	0.147	0.115	0.020	0.00%		0.00%		0.00%
master	        full	17.394	0.271	17.255	0.259	0.139	0.020	0.00%		0.00%		0.00%
master	        libs	15.758	0.343	15.613	0.329	0.145	0.021	0.00%		0.00%		0.00%
C2-combinator	cached	10.603	0.136	10.492	0.126	0.111	0.013	1.17%		1.14%		3.98%
C2-combinator	full	17.336	0.242	17.199	0.230	0.137	0.021	0.33%		0.33%		0.89%
C2-combinator	libs	15.658	0.204	15.518	0.190	0.140	0.022	0.63%		0.61%		3.34%
C2-combinator	cached	11.301	11.104	0.197	outlier